add search

+++ psLib.ps1
@@ -342,11 +342,6 @@ function recentlyChangeFile ([int]$weekCnt) {
     Get-ChildItem | Where-Object {$_.CreationTime -gt (Get-Date).AddDays($weekCnt * (-7))}
 }
 
-function logPersonalAlias () {
-    # only log can not work
-    Set-Alias rcf recentlyChangeFile;  
-}
-
 function eld {
     # enter lastAccessDirectory
     $lastAccessTimeDirectory = (Get-ChildItem -Directory . | Sort-Object LastAccessTime -Descending | Select-Object -First 1)
@@ -963,8 +958,23 @@ function zCleanCachedAndKillNode {
 }
 
 function zWorkCleanCachedAndKillNode {
-    Stop-Process -Name node 
-    $cachedFilePath = 'E:\DataIn\WorkFor\customerManage\consultantMobile\node_modules\.cache\hard-source\a46b2c3f7582d91717f14d4898282db2a8c6f33499ff80145e10b7a50c53e255'
+    # danger
+    $ErrorActionPreference = "SilentlyContinue"
+        
+    $PgNodeCnt = (Get-Process | Measure-Object).Count
+    if ($PgNodeCnt -ne 0) {
+        Stop-Process -Name node 
+        $cachedFilePath = 'E:\DataIn\WorkFor\customerManage\consultantMobile\node_modules\.cache\hard-source\a46b2c3f7582d91717f14d4898282db2a8c6f33499ff80145e10b7a50c53e255'
 
-    Remove-Item -Recurse $cachedFilePath
+        Remove-Item -Recurse $cachedFilePath
+
+    }
+    else {
+        Write-Host "\nno process node"
+    }
+    $ErrorActionPreference = "Continue"
+}
+
+function ToNegitiveNum ($param1 = $(Throw "param 1 must be specified!")) {
+    0 - [System.Convert]::ToInt32($param1)
 }
